Persist negotiated cipher suite
Extract the negotiated cipher suite from the TLS connection and restore
it when resuming the connection.

Adjusted tests so when we test the client we add an explicit cipher
suite to the server and vice versa. This way we can test that the
negotiated cipher suite is the one we expect.